Add AllocatePendingSession method (in draft phase);
Minor cleanup / code migration;

+++ src/HelpDesc.Core/TeamConfig.cs
@@ -7,7 +7,7 @@ public record TeamsConfig
 {
     public int MaximumConcurrency { get; set; }
 
-    public int MaximumQueueCapacityMultiplier { get; set; }
+    public double MaximumQueueCapacityMultiplier { get; set; }
 
     public List<SeniorityDescription> SeniorityDescriptions { get; set; }
